extracted the hook to its own file for clarity
