+++ src/Components/signIn/validate.js
@@ -1,13 +1,5 @@
 const validateAuth = values => {
   let errors = {};
-
-  var Repository = [
-    {email: 'drumac2@mail.com', password: 'aaaaaaaaa'},
-    {email: 'a@mail.com', password: 'aaaaaaaaa'},
-    {email: 'few@mail.com', password: 'aaaaaaaaa'},
-  ];
-
-  const result = Repository.find(user => user.email === values.email && user.password === values.password);
   
   // Email Errors
   if (!values.email) {
@@ -21,11 +13,6 @@ const validateAuth = values => {
   if (!values.password) {
     errors.password = 'Password is Required';
   } 
-
-  // Login
-  if (!result) {
-    errors.loggingIn = 'User with this email is not found';
-  }
 
   return errors;
 };
