+++ src/Reducer/authReducer.js
@@ -1,7 +1,9 @@
 function authReducer(state, action) {
   console.log(state, action);
+
   switch (action.type) {
     case 'LOGIN':
+      // HERE WE CALL GRAPHQL 
       return {
         ...state,
         user: action.payload,
